* WinApi osztály külön fájlba helyezve.
